+++ Libplanet.Tests/Blockchain/Renderers/AnonymousActionRendererTest.cs
@@ -27,6 +27,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
                 default,
                 default,
                 default,
+                Block.CurrentProtocolVersion,
                 default,
                 _stateDelta,
                 default,